pyproject.toml
[project]
name = "tg-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cerebras-cloud-sdk>=1.35.0",
    "telethon>=1.40.0",
    "python-dotenv>=1.0.0",
]
